test(material/slider): clean up harness testing setup
Now that we know that the MDC-based slider won't have the same API as the non-MDC version, we can avoid sharing the test harness tests and clean up some of the logic.
