+++ src/material-experimental/mdc-slider/testing/BUILD.bazel
@@ -24,7 +24,6 @@ ng_test_library(
     deps = [
         ":testing",
         "//src/material-experimental/mdc-slider",
-        "//src/material/slider/testing:harness_tests_lib",
     ],
 )
